Fix AtkUnitList.Count

+++ FFXIVClientStructs/FFXIV/Component/GUI/AtkUnitList.cs
@@ -8,5 +8,5 @@ namespace FFXIVClientStructs.FFXIV.Component.GUI;
 public unsafe struct AtkUnitList {
     [FieldOffset(0x0)] public void* vtbl;
     [FieldOffset(0x8)] public AtkUnitBase* AtkUnitEntries; // array of pointers 0x8-0x808
-    [FieldOffset(0x808)] public uint Count;
+    [FieldOffset(0x808)] public ushort Count;
 }
